Use 'hyperref' package for hyperlink support
In hyperref, the \url{...} command requires a protocol prefix, e.g.
"https://". We don't want this in MLA format, so the \url{...} command
has been redefined to add this "https://" prefix to the HREF but not the
displayed link.

This means that all links passed to \url{...} MUST NOT have a protocol
prefix.

This will work:

	\url{www.example.com/example-article.html}

This will fail:

	\url{https://www.example.com/example-article.html}